feat: port elm test from E2E to integration
