+++ src/modules/elm.rs
@@ -39,3 +39,62 @@ fn format_elm_version(elm_version: &str) -> Option<String> {
     let formatted_version = format!("v{}", elm_version.trim());
     Some(formatted_version)
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::modules::utils::test::render_module;
+    use ansi_term::Color;
+    use std::fs::{self, File};
+    use std::io;
+    use tempfile;
+
+    #[test]
+    fn folder_without_elm() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        let actual = render_module("elm", dir.path());
+        let expected = None;
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_elm_json() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("elm.json"))?.sync_all()?;
+        let actual = render_module("elm", dir.path());
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🌳 v0.19.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_elm_package_json() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("elm-package.json"))?.sync_all()?;
+        let actual = render_module("elm", dir.path());
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🌳 v0.19.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_elm_stuff_directory() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        let elmstuff = dir.path().join("elm-stuff");
+        fs::create_dir_all(&elmstuff)?;
+        let actual = render_module("elm", dir.path());
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🌳 v0.19.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+
+    #[test]
+    fn folder_with_elm_file() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("main.elm"))?.sync_all()?;
+        let actual = render_module("elm", dir.path());
+        let expected = Some(format!("via {} ", Color::Cyan.bold().paint("🌳 v0.19.1")));
+        assert_eq!(expected, actual);
+        Ok(())
+    }
+}
