+++ src/utils.rs
@@ -37,6 +37,10 @@ pub fn exec_cmd(cmd: &str, args: &[&str]) -> Option<CommandOutput> {
         _ => format!("{} {}", cmd, args.join(" ")),
     };
     match command.as_str() {
+        "elm --version" => Some(CommandOutput {
+            stdout: String::from("0.19.1"),
+            stderr: String::default(),
+        }),
         "node --version" => Some(CommandOutput {
             stdout: String::from("v12.0.0"),
             stderr: String::default(),
